backend/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
fastembed
biopython
chonkie
numpy
python-dotenv
